H5 - Adicionando Correçoes do servico e testes

+++ insurances/src/main/java/com/audsat/insurances/controller/InsuranceBudgetProposalController.java
@@ -1,8 +1,10 @@
 package com.audsat.insurances.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,7 +12,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.audsat.insurances.bean.InsuranceBudgetProposalBean;
 import com.audsat.insurances.model.Car;
@@ -47,7 +51,7 @@ public class InsuranceBudgetProposalController {
 		Insurance insuranceBudgetProposal = insuranceBudgetProposalService.getInsuranceBudgetProposalById(id);
 
 		if (insuranceBudgetProposal == null) {
-			return ResponseEntity.notFound().build();
+		     throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Insurance not found");
 		} else {
 			return ResponseEntity.ok(insuranceBudgetProposal);
 		}
@@ -57,14 +61,15 @@ public class InsuranceBudgetProposalController {
 	public ResponseEntity<List<Insurance>> findByAll() {
 		List<Insurance> insuranceBudgetProposal = insuranceBudgetProposalService.getAllInsuranceBudgetProposals();
 
-		if (insuranceBudgetProposal == null) {
-			return ResponseEntity.notFound().build();
+		if (insuranceBudgetProposal == null || insuranceBudgetProposal.isEmpty()) {
+		     throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Insurances not found");
 		} else {
 			return ResponseEntity.ok(insuranceBudgetProposal);
 		}
 	}
 
 	@PostMapping
+	@ResponseStatus(HttpStatus.CREATED)
     public ResponseEntity<Insurance> create(@RequestBody InsuranceBudgetProposalRequest insuranceBudgetProposalRequest) {
     	
     	
@@ -75,17 +80,25 @@ public class InsuranceBudgetProposalController {
     }
 
 	@DeleteMapping("/{id}")
+	@ResponseStatus(HttpStatus.NO_CONTENT)
 	public ResponseEntity<Void> deleteById(@PathVariable long id) {
-		insuranceBudgetProposalService.deleteInsuranceBudgetProposalById(id);
+		
+		Insurance insuranceBudgetProposal = insuranceBudgetProposalService.getInsuranceBudgetProposalById(id);
+
+	    if (insuranceBudgetProposal == null) {
+	        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Insurance not found");
+	    }
+	    
+	    insuranceBudgetProposalService.deleteInsuranceBudgetProposalById(id);
 		return ResponseEntity.ok().build();
 	}
 	
-	public InsuranceBudgetProposalBean parsetoBean(InsuranceBudgetProposalRequest request) {
+	private InsuranceBudgetProposalBean parsetoBean(InsuranceBudgetProposalRequest request) {
 		  InsuranceBudgetProposalBean bean = new InsuranceBudgetProposalBean();
 		  
 		  Customer customer = customerService.getCustomerBycpf(request.getCustomerDocument());
 		  
-		  Car car = carService.getCarByPlaca(request.getPlaca());
+		  Optional<Car> car = carService.getCarByPlaca(request.getPlaca());
 
 		  bean.setCustomer(customer); 
 		  bean.setCreateDt(request.getCreateDt());
